extract StaticMap class from static_map.php
